+++ metaspace/webapp/src/modules/GroupProfile/TransferDatasetsDialog.spec.ts
@@ -1,10 +1,9 @@
-import { mount, Wrapper } from '@vue/test-utils';
-import VueRouter from 'vue-router';
+import { mount, Stubs, Wrapper } from '@vue/test-utils';
 import ElementUI from 'element-ui';
 import Vue from 'vue';
 import TransferDatasetsDialog from './TransferDatasetsDialog.vue';
 import router from '../../router';
-jest.mock('../../components/DatasetItem.vue', () => require('../../../tests/utils/mockComponent')('dataset-item'));
+// jest.mock('../../components/DatasetItem.vue', () => require('../../../tests/utils/mockComponent')('dataset-item'));
 
 describe('TransferDatasetsDialog', () => {
   const mockDatasets = [
@@ -18,11 +17,14 @@ describe('TransferDatasetsDialog', () => {
     groupName: 'Group Name',
     isInvited: true
   };
+  const stubs: Stubs = {
+    DatasetItem: true
+  };
   [false, true].forEach(hasDatasets => {
     [false, true].forEach(isInvited => {
       it(`should match snapshot (${hasDatasets ? 'datasets to import' : 'no datasets'}, ${isInvited ? 'invited' : 'requesting access'})`, () => {
         const propsData = { ...mockProps, isInvited };
-        const wrapper = mount(TransferDatasetsDialog, { router, propsData, sync: false });
+        const wrapper = mount(TransferDatasetsDialog, { router, propsData, stubs, sync: false });
         wrapper.setData({ allDatasets: hasDatasets ? mockDatasets : [] });
 
         expect(wrapper).toMatchSnapshot();
@@ -31,7 +33,7 @@ describe('TransferDatasetsDialog', () => {
   });
 
   it('should call back on success when some datasets are selected', async () => {
-    const wrapper = mount(TransferDatasetsDialog, { router, propsData: mockProps, sync: false });
+    const wrapper = mount(TransferDatasetsDialog, { router, propsData: mockProps, stubs, sync: false });
     wrapper.setData({ allDatasets: mockDatasets });
     await Vue.nextTick();
