+++ metaspace/webapp/src/modules/GroupProfile/ViewGroupProfile.vue
@@ -67,7 +67,7 @@
   import Vue from 'vue';
   import { Component } from 'vue-property-decorator';
   import { DatasetDetailItem, datasetDetailItemFragment } from '../../api/dataset';
-  import DatasetList from '../../components/DatasetList.vue';
+  import DatasetList from '../Datasets/list/DatasetList.vue';
   import {
     acceptGroupInvitationMutation,
     importDatasetsIntoGroupMutation,
