+++ metaspace/webapp/src/modules/Project/EditProjectPage.vue
@@ -55,7 +55,6 @@
 <script lang="ts">
   import Vue from 'vue';
   import { Component, Watch } from 'vue-property-decorator';
-  import DatasetItem from '../../components/DatasetItem.vue';
   import {
     acceptRequestToJoinProjectMutation,
     deleteProjectMutation,
@@ -83,7 +82,6 @@
 
   @Component({
     components: {
-      DatasetItem,
       EditProjectForm,
       MembersList,
     },
